backend/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.7.0
Pillow==10.4.0
python-decouple==3.8
django-filter==25.1
